pyproject.toml
[project]
name = "merry-lights-demos"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "marimo>=0.9.27",
    "pillow>=11.0.0",
]
